+++ src/app.module.ts
@@ -23,8 +23,8 @@ import { DataSource } from 'typeorm';
 import * as console from 'node:console';
 import { addTransactionalDataSource } from 'typeorm-transactional';
 import { EntitySubscriber } from './subscriber/entity.subscriber';
-import { JwtAuthGuard } from './auth/guards/jwt-auth.guard';
-import { RolesGuard } from './auth/guards/roles.guard';
+import { JwtAuthGuard } from './common/guards/jwt-auth.guard';
+import { RolesGuard } from './common/guards/roles.guard';
 import auth20SecretConfig from './config/auth20-secret.config';
 import fileUploadConfig from './config/file-upload.config';
 import { InterestModule } from './interest/interest.module';
