+++ src/interest/interest.controller.ts
@@ -2,7 +2,7 @@ import { Body, Controller, Get, Post } from '@nestjs/common';
 import { ApiBearerAuth, ApiExtraModels, ApiTags } from '@nestjs/swagger';
 import { AllInterestResDto, InterestResDto } from './dto/res.dto';
 import { Roles } from '../common/decorater/roles.decorator';
-import { Role } from '../user/common/enums/role.enum';
+import { Role } from '../user/enums/role.enum';
 import { InterestService } from './interest.service';
 import { InterestReqDto } from './dto/req.dto';
 import { ApiGetResponse } from '../common/decorater/swagger.decorator';
