+++ src/tech-stack/dto/res.dto.ts
@@ -1,5 +1,5 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { Role } from '../../user/common/enums/role.enum';
+import { Role } from '../../user/enums/role.enum';
 import { TechStack } from '../entity/tech-stack.entity';
 
 export class TechStackResDto {
